refactor: moves messaging styles to messaging component file
Relates #422
